feat - added post page, to view single posts
